get rid of the meta meta strategy
